@@ -19,13 +19,14 @@ import android.util.Log;
 
 import androidx.core.view.ViewCompat;
 
-import com.google.android.exoplayer2.text.CaptionStyleCompat;
+import com.google.android.exoplayer2.ui.CaptionStyleCompat;
 import com.google.android.exoplayer2.text.Cue;
 import com.google.android.exoplayer2.util.Util;
 
-public class SubtitlePainter  {
+public class SubtitlePainter {
     private static final float INNER_PADDING_RATIO = 0.125f;
     private static final String TAG = "SubtitlePainter";
+
     private boolean applyEmbeddedFontSizes;
     private boolean applyEmbeddedStyles;
     private int backgroundColor;
@@ -65,18 +66,22 @@ public class SubtitlePainter  {
     private int windowColor;
 
     public SubtitlePainter(Context context) {
-        TypedArray styledAttributes = context.obtainStyledAttributes(null, new int[]{16843287, 16843288}, 0, 0);
-        this.spacingAdd = (float) styledAttributes.getDimensionPixelSize(0, 0);
+        TypedArray styledAttributes = context.obtainStyledAttributes(
+                null, new int[]{android.R.attr.lineSpacingExtra, android.R.attr.lineSpacingMultiplier}, 0, 0);
+        this.spacingAdd = styledAttributes.getDimensionPixelSize(0, 0);
         this.spacingMult = styledAttributes.getFloat(1, 1.0f);
         styledAttributes.recycle();
-        int twoDpInPx = Math.round((2.0f * ((float) context.getResources().getDisplayMetrics().densityDpi)) / 160.0f);
-        this.cornerRadius = (float) twoDpInPx;
-        this.outlineWidth = (float) twoDpInPx;
-        this.shadowRadius = (float) twoDpInPx;
-        this.shadowOffset = (float) twoDpInPx;
+
+        int twoDpInPx = Math.round(2.0f * (context.getResources().getDisplayMetrics().densityDpi / 160f));
+        this.cornerRadius = twoDpInPx;
+        this.outlineWidth = twoDpInPx;
+        this.shadowRadius = twoDpInPx;
+        this.shadowOffset = twoDpInPx;
+
         this.textPaint = new TextPaint();
         this.textPaint.setAntiAlias(true);
         this.textPaint.setSubpixelText(true);
+
         this.paint = new Paint();
         this.paint.setAntiAlias(true);
         this.paint.setStyle(Paint.Style.FILL);
@@ -86,9 +91,22 @@ public class SubtitlePainter  {
         return first == second || (first != null && first.equals(second));
     }
 
-    public void draw(Cue cue, boolean applyEmbeddedStyles, boolean applyEmbeddedFontSizes, CaptionStyleCompat style, float textSizePx, float bottomPaddingFraction, Canvas canvas, int cueBoxLeft, int cueBoxTop, int cueBoxRight, int cueBoxBottom) {
+    public void draw(
+            Cue cue,
+            boolean applyEmbeddedStyles,
+            boolean applyEmbeddedFontSizes,
+            CaptionStyleCompat style,
+            float textSizePx,
+            float bottomPaddingFraction,
+            Canvas canvas,
+            int cueBoxLeft,
+            int cueBoxTop,
+            int cueBoxRight,
+            int cueBoxBottom
+    ) {
         boolean isTextCue = cue.bitmap == null;
         int windowColor = ViewCompat.MEASURED_STATE_MASK;
+
         if (isTextCue) {
             if (!TextUtils.isEmpty(cue.text)) {
                 windowColor = (cue.windowColorSet && applyEmbeddedStyles) ? cue.windowColor : style.windowColor;
@@ -96,10 +114,35 @@ public class SubtitlePainter  {
                 return;
             }
         }
-        if (areCharSequencesEqual(this.cueText, cue.text) && Util.areEqual(this.cueTextAlignment, cue.textAlignment) && this.cueBitmap == cue.bitmap && this.cueLine == cue.line && this.cueLineType == cue.lineType && Util.areEqual(Integer.valueOf(this.cueLineAnchor), Integer.valueOf(cue.lineAnchor)) && this.cuePosition == cue.position && Util.areEqual(Integer.valueOf(this.cuePositionAnchor), Integer.valueOf(cue.positionAnchor)) && this.cueSize == cue.size && this.cueBitmapHeight == cue.bitmapHeight && this.applyEmbeddedStyles == applyEmbeddedStyles && this.applyEmbeddedFontSizes == applyEmbeddedFontSizes && this.foregroundColor == style.foregroundColor && this.backgroundColor == style.backgroundColor && this.windowColor == windowColor && this.edgeType == style.edgeType && this.edgeColor == style.edgeColor && Util.areEqual(this.textPaint.getTypeface(), style.typeface) && this.textSizePx == textSizePx && this.bottomPaddingFraction == bottomPaddingFraction && this.parentLeft == cueBoxLeft && this.parentTop == cueBoxTop && this.parentRight == cueBoxRight && this.parentBottom == cueBoxBottom) {
+
+        if (areCharSequencesEqual(this.cueText, cue.text)
+                && Util.areEqual(this.cueTextAlignment, cue.textAlignment)
+                && this.cueBitmap == cue.bitmap
+                && this.cueLine == cue.line
+                && this.cueLineType == cue.lineType
+                && Util.areEqual(this.cueLineAnchor, cue.lineAnchor)
+                && this.cuePosition == cue.position
+                && Util.areEqual(this.cuePositionAnchor, cue.positionAnchor)
+                && this.cueSize == cue.size
+                && this.cueBitmapHeight == cue.bitmapHeight
+                && this.applyEmbeddedStyles == applyEmbeddedStyles
+                && this.applyEmbeddedFontSizes == applyEmbeddedFontSizes
+                && this.foregroundColor == style.foregroundColor
+                && this.backgroundColor == style.backgroundColor
+                && this.windowColor == windowColor
+                && this.edgeType == style.edgeType
+                && this.edgeColor == style.edgeColor
+                && Util.areEqual(this.textPaint.getTypeface(), style.typeface)
+                && this.textSizePx == textSizePx
+                && this.bottomPaddingFraction == bottomPaddingFraction
+                && this.parentLeft == cueBoxLeft
+                && this.parentTop == cueBoxTop
+                && this.parentRight == cueBoxRight
+                && this.parentBottom == cueBoxBottom) {
             drawLayout(canvas, isTextCue);
             return;
         }
+
         this.cueText = cue.text;
         this.cueTextAlignment = cue.textAlignment;
         this.cueBitmap = cue.bitmap;
@@ -124,6 +167,7 @@ public class SubtitlePainter  {
         this.parentTop = cueBoxTop;
         this.parentRight = cueBoxRight;
         this.parentBottom = cueBoxBottom;
+
         if (isTextCue) {
             setupTextLayout();
         } else {
@@ -135,113 +179,139 @@ public class SubtitlePainter  {
     private void setupTextLayout() {
         int parentWidth = this.parentRight - this.parentLeft;
         int parentHeight = this.parentBottom - this.parentTop;
+
         this.textPaint.setTextSize(this.textSizePx);
-        int textPaddingX = (int) ((this.textSizePx * INNER_PADDING_RATIO) + 0.5f);
+        int textPaddingX = (int) (this.textSizePx * INNER_PADDING_RATIO + 0.5f);
         int availableWidth = parentWidth - (textPaddingX * 2);
         if (this.cueSize != Float.MIN_VALUE) {
-            availableWidth = (int) (((float) availableWidth) * this.cueSize);
+            availableWidth = (int) (availableWidth * this.cueSize);
         }
         if (availableWidth <= 0) {
             Log.w(TAG, "Skipped drawing subtitle cue (insufficient space)");
             return;
         }
-        CharSequence cueText = null;
-        int textLeft;
-        int textRight;
+
+        CharSequence cueText;
         if (this.applyEmbeddedFontSizes && this.applyEmbeddedStyles) {
             cueText = this.cueText;
         } else if (this.applyEmbeddedStyles) {
-            SpannableStringBuilder spannableStringBuilder = new SpannableStringBuilder(this.cueText);
-            int cueLength = spannableStringBuilder.length();
-            RelativeSizeSpan[] relSpans = (RelativeSizeSpan[]) spannableStringBuilder.getSpans(0, cueLength, RelativeSizeSpan.class);
-            for (AbsoluteSizeSpan absSpan : (AbsoluteSizeSpan[]) spannableStringBuilder.getSpans(0, cueLength, AbsoluteSizeSpan.class)) {
-                spannableStringBuilder.removeSpan(absSpan);
+            SpannableStringBuilder ssb = new SpannableStringBuilder(this.cueText);
+            int cueLength = ssb.length();
+            for (AbsoluteSizeSpan absSpan : ssb.getSpans(0, cueLength, AbsoluteSizeSpan.class)) {
+                ssb.removeSpan(absSpan);
             }
-            for (RelativeSizeSpan relSpan : relSpans) {
-                spannableStringBuilder.removeSpan(relSpan);
+            for (RelativeSizeSpan relSpan : ssb.getSpans(0, cueLength, RelativeSizeSpan.class)) {
+                ssb.removeSpan(relSpan);
             }
-            Object cueText2 = spannableStringBuilder;
+            cueText = ssb; // <— ensure we keep styled text but strip font sizes
         } else {
             cueText = this.cueText.toString();
         }
-        Layout.Alignment textAlignment = this.cueTextAlignment == null ? Layout.Alignment.ALIGN_CENTER : this.cueTextAlignment;
-        this.textLayout = new StaticLayout(cueText, this.textPaint, availableWidth, textAlignment, this.spacingMult, this.spacingAdd, true);
+
+        Layout.Alignment textAlignment =
+                this.cueTextAlignment == null ? Layout.Alignment.ALIGN_CENTER : this.cueTextAlignment;
+
+        this.textLayout = new StaticLayout(
+                cueText, this.textPaint, availableWidth, textAlignment, this.spacingMult, this.spacingAdd, true);
+
         int textHeight = this.textLayout.getHeight();
         int textWidth = 0;
         int lineCount = this.textLayout.getLineCount();
         for (int i = 0; i < lineCount; i++) {
-            textWidth = Math.max((int) Math.ceil((double) this.textLayout.getLineWidth(i)), textWidth);
+            textWidth = Math.max((int) Math.ceil(this.textLayout.getLineWidth(i)), textWidth);
         }
         if (this.cueSize != Float.MIN_VALUE && textWidth < availableWidth) {
             textWidth = availableWidth;
         }
         textWidth += textPaddingX * 2;
+
+        int textLeft;
+        int textRight;
         if (this.cuePosition != Float.MIN_VALUE) {
-            int anchorPosition = Math.round(((float) parentWidth) * this.cuePosition) + this.parentLeft;
-            textLeft = this.cuePositionAnchor == 2 ? anchorPosition - textWidth : this.cuePositionAnchor == 1 ? ((anchorPosition * 2) - textWidth) / 2 : anchorPosition;
+            int anchorPosition = Math.round(parentWidth * this.cuePosition) + this.parentLeft;
+            if (this.cuePositionAnchor == 2) {
+                textLeft = anchorPosition - textWidth;
+            } else if (this.cuePositionAnchor == 1) {
+                textLeft = ((anchorPosition * 2) - textWidth) / 2;
+            } else {
+                textLeft = anchorPosition;
+            }
             textLeft = Math.max(textLeft, this.parentLeft);
             textRight = Math.min(textLeft + textWidth, this.parentRight);
         } else {
             textLeft = (parentWidth - textWidth) / 2;
             textRight = textLeft + textWidth;
         }
+
         textWidth = textRight - textLeft;
         if (textWidth <= 0) {
             Log.w(TAG, "Skipped drawing subtitle cue (invalid horizontal positioning)");
             return;
         }
+
         int textTop;
         if (this.cueLine != Float.MIN_VALUE) {
             int anchorPosition;
             if (this.cueLineType == 0) {
-                anchorPosition = Math.round(((float) parentHeight) * this.cueLine) + this.parentTop;
+                anchorPosition = Math.round(parentHeight * this.cueLine) + this.parentTop;
             } else {
                 int firstLineHeight = this.textLayout.getLineBottom(0) - this.textLayout.getLineTop(0);
                 if (this.cueLine >= 0.0f) {
-                    anchorPosition = Math.round(this.cueLine * ((float) firstLineHeight)) + this.parentTop;
+                    anchorPosition = Math.round(this.cueLine * firstLineHeight) + this.parentTop;
                 } else {
-                    anchorPosition = Math.round((this.cueLine + 1.0f) * ((float) firstLineHeight)) + this.parentBottom;
+                    anchorPosition = Math.round((this.cueLine + 1.0f) * firstLineHeight) + this.parentBottom;
                 }
             }
-            textTop = this.cueLineAnchor == 2 ? anchorPosition - textHeight : this.cueLineAnchor == 1 ? ((anchorPosition * 2) - textHeight) / 2 : anchorPosition;
+            if (this.cueLineAnchor == 2) {
+                textTop = anchorPosition - textHeight;
+            } else if (this.cueLineAnchor == 1) {
+                textTop = ((anchorPosition * 2) - textHeight) / 2;
+            } else {
+                textTop = anchorPosition;
+            }
             if (textTop + textHeight > this.parentBottom) {
                 textTop = this.parentBottom - textHeight;
             } else if (textTop < this.parentTop) {
                 textTop = this.parentTop;
             }
         } else {
-            textTop = (this.parentBottom - textHeight) - ((int) (((float) parentHeight) * this.bottomPaddingFraction));
+            textTop = this.parentBottom - textHeight - (int) (parentHeight * this.bottomPaddingFraction);
         }
-        this.textLayout = new StaticLayout(cueText, this.textPaint, textWidth, textAlignment, this.spacingMult, this.spacingAdd, true);
+
+        this.textLayout = new StaticLayout(
+                cueText, this.textPaint, textWidth, textAlignment, this.spacingMult, this.spacingAdd, true);
+
         this.textLeft = textLeft;
         this.textTop = textTop;
         this.textPaddingX = textPaddingX;
     }
 
     private void setupBitmapLayout() {
-        int height;
         int parentWidth = this.parentRight - this.parentLeft;
         int parentHeight = this.parentBottom - this.parentTop;
-        float anchorX = ((float) this.parentLeft) + (((float) parentWidth) * this.cuePosition);
-        float anchorY = ((float) this.parentTop) + (((float) parentHeight) * this.cueLine);
-        int width = Math.round(((float) parentWidth) * this.cueSize);
-        if (this.cueBitmapHeight != Float.MIN_VALUE) {
-            height = Math.round(((float) parentHeight) * this.cueBitmapHeight);
-        } else {
-            height = Math.round(((float) width) * (((float) this.cueBitmap.getHeight()) / ((float) this.cueBitmap.getWidth())));
-        }
+
+        float anchorX = this.parentLeft + parentWidth * this.cuePosition;
+        float anchorY = this.parentTop + parentHeight * this.cueLine;
+
+        int width = Math.round(parentWidth * this.cueSize);
+        int height = (this.cueBitmapHeight != Float.MIN_VALUE)
+                ? Math.round(parentHeight * this.cueBitmapHeight)
+                : Math.round(width * (this.cueBitmap.getHeight() / (float) this.cueBitmap.getWidth()));
+
         if (this.cueLineAnchor == 2) {
-            anchorX -= (float) width;
+            anchorX -= width;
         } else if (this.cueLineAnchor == 1) {
-            anchorX -= (float) (width / 2);
+            anchorX -= (width / 2f);
         }
         int x = Math.round(anchorX);
+
         if (this.cuePositionAnchor == 2) {
-            anchorY -= (float) height;
+            anchorY -= height;
         } else if (this.cuePositionAnchor == 1) {
-            anchorY -= (float) (height / 2);
+            anchorY -= (height / 2f);
         }
         int y = Math.round(anchorY);
+
         this.bitmapRect = new Rect(x, y, x + width, y + height);
     }
 
@@ -255,51 +325,58 @@ public class SubtitlePainter  {
 
     private void drawTextLayout(Canvas canvas) {
         StaticLayout layout = this.textLayout;
-        if (layout != null) {
-            int saveCount = canvas.save();
-            canvas.translate((float) this.textLeft, (float) this.textTop);
-            if (Color.alpha(this.windowColor) > 0) {
-                this.paint.setColor(this.windowColor);
-                canvas.drawRect((float) (-this.textPaddingX), 0.0f, (float) (layout.getWidth() + this.textPaddingX), (float) layout.getHeight(), this.paint);
+        if (layout == null) return;
+
+        int save = canvas.save();
+        canvas.translate(this.textLeft, this.textTop);
+
+        if (Color.alpha(this.windowColor) > 0) {
+            this.paint.setColor(this.windowColor);
+            canvas.drawRect(-this.textPaddingX, 0, layout.getWidth() + this.textPaddingX, layout.getHeight(), this.paint);
+        }
+
+        if (Color.alpha(this.backgroundColor) > 0) {
+            this.paint.setColor(this.backgroundColor);
+            float previousBottom = layout.getLineTop(0);
+            int lineCount = layout.getLineCount();
+            for (int i = 0; i < lineCount; i++) {
+                this.lineBounds.left = layout.getLineLeft(i) - this.textPaddingX;
+                this.lineBounds.right = layout.getLineRight(i) + this.textPaddingX;
+                this.lineBounds.top = previousBottom;
+                this.lineBounds.bottom = layout.getLineBottom(i);
+                previousBottom = this.lineBounds.bottom;
+                canvas.drawRoundRect(this.lineBounds, this.cornerRadius, this.cornerRadius, this.paint);
             }
-            if (Color.alpha(this.backgroundColor) > 0) {
-                this.paint.setColor(this.backgroundColor);
-                float previousBottom = (float) layout.getLineTop(0);
-                int lineCount = layout.getLineCount();
-                for (int i = 0; i < lineCount; i++) {
-                    this.lineBounds.left = layout.getLineLeft(i) - ((float) this.textPaddingX);
-                    this.lineBounds.right = layout.getLineRight(i) + ((float) this.textPaddingX);
-                    this.lineBounds.top = previousBottom;
-                    this.lineBounds.bottom = (float) layout.getLineBottom(i);
-                    previousBottom = this.lineBounds.bottom;
-                    canvas.drawRoundRect(this.lineBounds, this.cornerRadius, this.cornerRadius, this.paint);
-                }
-            }
-            if (this.edgeType == 1) {
-                this.textPaint.setStrokeJoin(Paint.Join.ROUND);
-                this.textPaint.setStrokeWidth(this.outlineWidth);
-                this.textPaint.setColor(this.edgeColor);
-                this.textPaint.setStyle(Paint.Style.FILL_AND_STROKE);
-                layout.draw(canvas);
-            } else if (this.edgeType == 2) {
-                this.textPaint.setShadowLayer(this.shadowRadius, this.shadowOffset, this.shadowOffset, this.edgeColor);
-            } else if (this.edgeType == 3 || this.edgeType == 4) {
-                boolean raised = this.edgeType == 3;
-                int colorUp = raised ? -1 : this.edgeColor;
-                int colorDown = raised ? this.edgeColor : -1;
-                float offset = this.shadowRadius / 2.0f;
-                this.textPaint.setColor(this.foregroundColor);
-                this.textPaint.setStyle(Paint.Style.FILL);
-                this.textPaint.setShadowLayer(this.shadowRadius, -offset, -offset, colorUp);
-                layout.draw(canvas);
-                this.textPaint.setShadowLayer(this.shadowRadius, offset, offset, colorDown);
-            }
+        }
+
+        if (this.edgeType == 1) { // outline
+            this.textPaint.setStrokeJoin(Paint.Join.ROUND);
+            this.textPaint.setStrokeWidth(this.outlineWidth);
+            this.textPaint.setColor(this.edgeColor);
+            this.textPaint.setStyle(Paint.Style.FILL_AND_STROKE);
+            layout.draw(canvas);
+        } else if (this.edgeType == 2) { // drop shadow
+            this.textPaint.setShadowLayer(this.shadowRadius, this.shadowOffset, this.shadowOffset, this.edgeColor);
+        } else if (this.edgeType == 3 || this.edgeType == 4) { // raised/depressed
+            boolean raised = this.edgeType == 3;
+            int colorUp = raised ? Color.WHITE : this.edgeColor;
+            int colorDown = raised ? this.edgeColor : Color.WHITE;
+            float offset = this.shadowRadius / 2f;
+
             this.textPaint.setColor(this.foregroundColor);
             this.textPaint.setStyle(Paint.Style.FILL);
+            this.textPaint.setShadowLayer(this.shadowRadius, -offset, -offset, colorUp);
             layout.draw(canvas);
-            this.textPaint.setShadowLayer(0.0f, 0.0f, 0.0f, 0);
-            canvas.restoreToCount(saveCount);
+
+            this.textPaint.setShadowLayer(this.shadowRadius, offset, offset, colorDown);
         }
+
+        this.textPaint.setColor(this.foregroundColor);
+        this.textPaint.setStyle(Paint.Style.FILL);
+        layout.draw(canvas);
+
+        this.textPaint.setShadowLayer(0f, 0f, 0f, 0);
+        canvas.restoreToCount(save);
     }
 
     private void drawBitmapLayout(Canvas canvas) {
